Initial commit.
This is my initial commit for the Grephy project. I have a basic bootstrap template for the layout of the UI, if this is an issue please let me know. It is a free to use template.

This commit includes the basics:

Basic UI
File select
Alerts for testing
Beginnings of file parsing